requirements.txt
imageio==2.9.0
numpy==1.19.4
Pillow==10.3.0
pytorch_msssim==0.2.1
SimpleITK==2.3.1
torch==1.8.1+cu101
torchvision==0.7.0+cu101
tqdm==4.64.0
